Added Thread Model for Login Use Case
This Thread Model is associated with the following misuse case: Wagtail is sufficiently secure against SQL injection attacks. 

From reviewing the codebase, it looks like Wagtail uses Django's Authentication Middleware. It is a pretty basic Thread model, but because it is using Django's middleware I don't think there is much need to expand on that portion of it.