Updated Threat Model for Login Use Case
My understanding is the following url patter in __init__.py
url(r'^login/$', account.login, name='wagtailadmin_login')  maps the
login page 'login.html' with the login view in account.py. The form
action in login.html is <form action="{% url 'wagtailadmin_login' %}"
method="post" autocomplete="off" novalidate> so wagtailadmin_login will
invoke the account.login view (function).

The login view takes the request and if the user is not authenticated
(user is AnonymousUser) then it calls django.contrib.auth.view.login().
It looks as if Wagtail uses the default authentication system provided
by Django. So the Authentication middleware actually gets the user from
the ModelBackend (db.sqlite3) auth_user table. The user is returned as
User and not AnonymouseUser so is_authenticated is True. The reponse
data is returned to the client and the page is redirected to home.